Mob kill counter on screen and new script for global vars
The kill counters really simple but we can spice it up later. Global variable script will be very useful for allowing all scrips to read/write to same vars. In this I also realised that _ready() as a function name has a purpose with this engine and can't be changed to something else.